Modificando tipos de variables para q sean todas unsigned long int (por si las moscas) ... y arreglando un error que tenía (gracias Gaby por avisarme)..era una pavada al final ..tenía dos veces definida la misma variables ..y en un lado la seteaba en cero jaja...q sopencaaaaa

+++ src/business/stego/BMPStego.h
@@ -7,7 +7,7 @@ class BMPStego:public LSBStegoBusiness
 public:
 	BMPStego();
 	BMPStego(std::string filename);
-	virtual unsigned int getFreeSpace();
+	virtual unsigned long int getFreeSpace();
 	virtual ~BMPStego();
 };
